Update README.md
Oh yeah, that reminds me of the principle of "least punishing consequences". Though I think I had a better name for it. But it boils down to an optimization method that doesn't try to maximize rewards but minimize negative consequences. Which is probably just another path to the same. And it sounds a bit like the least-action principle.